Implements modularized system for use-cases
Using the new @use_case and @configurable system, it is now possible to
add further use-cases to the wintermute project, while having them
automatically inject all required configurations.
It also formalizes some parts of the code, such as LLM capabilities,
which are now specifically created.
This is still a point where some functionality is missing though, as the
capabilities are not yet properly passed to function calling LLMs, which
should be done in a future release.

Proper care was taken, to not change the semantics of the resulting
use-cases (such as local_privesc_linux) other than that the
configuration now works differently. If there are regressions then this
is not on purpose and should be reported.

This commit also contains various code-cleanup operations, which
happened through the re-work for the new system. However, especially
helper methods, and the contents of the privesc use-case have largely
stayed unchanged, other than adapting them to the new surrounding
infrastructure.